Futher book model refactoring
