+++ ReadingList/Models/Book.swift
@@ -36,7 +36,6 @@ class Book: NSManagedObject {
     @NSManaged var coverImage: Data?
     @NSManaged var notes: String?
     @NSManaged var languageCode: String? // ISO 639.1: two-digit language code
-
     @NSManaged var subjects: Set<Subject>
     @NSManaged private(set) var lists: Set<List>
 
@@ -73,18 +72,6 @@ class Book: NSManagedObject {
         set { safelySetPrimitiveValue(newValue, .sort) }
     }
 
-    convenience init(context: NSManagedObjectContext, readState: BookReadState) {
-        self.init(context: context)
-        self.readState = readState
-        if readState == .reading {
-            startedReading = Date()
-        }
-        if readState == .finished {
-            startedReading = Date()
-            finishedReading = Date()
-        }
-    }
-
     override func willSave() {
         super.willSave()
 
@@ -112,7 +99,7 @@ class Book: NSManagedObject {
         super.prepareForDeletion()
         for orphanedSubject in subjects.filter({ $0.books.count == 1 }) {
             orphanedSubject.delete()
-            print("orphaned subject \(orphanedSubject.name) deleted.")
+            os_log("Orphaned subject %{public}s deleted.", type: .info, orphanedSubject.name)
         }
     }
 }
@@ -139,21 +126,6 @@ extension Book {
         authors = searchResult.authors
         isbn13 = ISBN13(searchResult.isbn13)?.int
         self.coverImage = coverImage
-    }
-
-    private func populateAuthors(fromStrings authors: [String]) {
-        let authorNames: [(String?, String)] = authors.map {
-            if let range = $0.range(of: " ", options: .backwards) {
-                let firstNames = $0[..<range.upperBound].trimming()
-                let lastName = $0[range.lowerBound...].trimming()
-
-                return (firstNames: firstNames, lastName: lastName)
-            } else {
-                return (firstNames: nil, lastName: $0)
-            }
-        }
-
-        self.authors = authorNames.map { Author(lastName: $0.1, firstNames: $0.0) }
     }
 
     static func get(fromContext context: NSManagedObjectContext, googleBooksId: String? = nil, isbn: String? = nil) -> Book? {
@@ -202,7 +174,7 @@ extension Book {
         fetchRequest.predicate = NSPredicate.and([
             NSPredicate(format: "%K == %ld", #keyPath(Book.readState), BookReadState.toRead.rawValue),
             NSPredicate(format: "%K != nil", Book.Key.sort.rawValue)])
-        fetchRequest.sortDescriptors = [NSSortDescriptor(\Book.sort, ascending: !getMaximum)]
+        fetchRequest.sortDescriptors = [NSSortDescriptor(Book.Key.sort.rawValue, ascending: !getMaximum)]
         fetchRequest.returnsObjectsAsFaults = false
         return (try! context.fetch(fetchRequest)).first?.sort
     }
@@ -213,72 +185,6 @@ extension Book {
 
     static func minSort(fromContext context: NSManagedObjectContext) -> Int32? {
         return maximalSort(getMaximum: false, fromContext: context)
-    }
-
-    @objc func validateAuthors(_ value: AutoreleasingUnsafeMutablePointer<AnyObject?>) throws {
-        // nil authors property will be validated by the validation set on the model
-        guard let authors = value.pointee as? [Author] else { return }
-        if authors.isEmpty {
-            throw BookValidationError.noAuthors.NSError()
-        }
-    }
-
-    @objc func validateTitle(_ value: AutoreleasingUnsafeMutablePointer<AnyObject?>) throws {
-        // nil title property will be validated by the validation set on the model
-        guard let title = value.pointee as? String else { return }
-        if title.isEmptyOrWhitespace {
-            throw BookValidationError.missingTitle.NSError()
-        }
-    }
-
-    @objc func validateIsbn13(_ value: AutoreleasingUnsafeMutablePointer<AnyObject?>) throws {
-        guard let isbn13 = value.pointee as? Int64 else { return }
-        if !ISBN13.isValid(isbn13) {
-            throw BookValidationError.invalidIsbn.NSError()
-        }
-    }
-
-    @objc func validateLanguageCode(_ value: AutoreleasingUnsafeMutablePointer<AnyObject?>) throws {
-        guard let languageCode = value.pointee as? String else { return }
-        if Language.byIsoCode[languageCode] == nil {
-            throw BookValidationError.invalidLanguageCode.NSError()
-        }
-    }
-
-    override func validateForUpdate() throws {
-        try super.validateForUpdate()
-        try interPropertyValiatation()
-    }
-
-    override func validateForInsert() throws {
-        try super.validateForInsert()
-        try interPropertyValiatation()
-    }
-
-    func interPropertyValiatation() throws {
-        switch readState {
-        case .toRead:
-            if startedReading != nil || finishedReading != nil {
-                throw BookValidationError.invalidReadDates.NSError()
-            }
-        case .reading:
-            if startedReading == nil || finishedReading != nil {
-                throw BookValidationError.invalidReadDates.NSError()
-            }
-        case .finished:
-            if startedReading == nil || finishedReading == nil {
-                throw BookValidationError.invalidReadDates.NSError()
-            }
-        }
-        if readState != .reading && currentPage != nil {
-            throw BookValidationError.presentCurrentPage.NSError()
-        }
-        if googleBooksId == nil && manualBookId == nil {
-            throw BookValidationError.missingIdentifier.NSError()
-        }
-        if googleBooksId != nil && manualBookId != nil {
-            throw BookValidationError.conflictingIdentifiers.NSError()
-        }
     }
 
     func startReading() {
@@ -298,35 +204,5 @@ extension Book {
         readState = .finished
         currentPage = nil
         finishedReading = Date()
-    }
-}
-
-enum BookValidationError: Int {
-    case missingTitle = 1
-    case invalidIsbn = 2
-    case invalidReadDates = 3
-    case invalidLanguageCode = 4
-    case missingIdentifier = 5
-    case conflictingIdentifiers = 6
-    case noAuthors = 7
-    case presentCurrentPage = 8
-}
-
-extension BookValidationError {
-    var description: String {
-        switch self {
-        case .missingTitle: return "Title must be non-empty or whitespace"
-        case .invalidIsbn: return "Isbn13 must be a valid ISBN"
-        case .invalidReadDates: return "StartedReading and FinishedReading must align with ReadState"
-        case .invalidLanguageCode: return "LanguageCode must be an ISO-639.1 value"
-        case .conflictingIdentifiers: return "GoogleBooksId and ManualBooksId cannot both be non nil"
-        case .missingIdentifier: return "GoogleBooksId and ManualBooksId cannot both be nil"
-        case .noAuthors: return "Authors array must be non-empty"
-        case .presentCurrentPage: return "CurrentPage must be nil when not Currently Reading"
-        }
-    }
-
-    func NSError() -> NSError {
-        return Foundation.NSError(domain: "BookErrorDomain", code: self.rawValue, userInfo: [NSLocalizedDescriptionKey: self.description])
     }
 }
